backend/color_ex/requirements.txt
asgiref==3.6.0
attrs==22.2.0
click==8.1.3
cloudpickle==2.2.1
Django==3.2.18
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
doit==0.36.0
drf-spectacular==0.25.1
h11==0.14.0
importlib-metadata==6.0.0
inflection==0.5.1
jsonschema==4.17.3
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
uvicorn==0.20.0
Werkzeug==2.2.3
whitenoise==6.4.0
zipp==3.15.0
